+++ controller/v1/env_test.go
@@ -1,4 +1,4 @@
-package controller
+package v1
 
 import (
 	"testing"
