+++ controller/v1/hostname.go
@@ -1,4 +1,4 @@
-package controller
+package v1
 
 import (
 	"encoding/json"
