+++ controller/v1/index.go
@@ -1,4 +1,4 @@
-package controller
+package v1
 
 import (
 	"net/http"
